Move ExecuteTemplate to be examples specific
